# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Project ciph3r",
        "description": "Documentacion del API",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar todos los productos",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Lista de productos con sus monedas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Test Update22"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Description test update."
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1000.1
                                            },
                                            "currency_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "tax_cost": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 15.01
                                            },
                                            "manufacturing_cost": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 357.66
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-03-12T18:21:39.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-03-12T20:06:33.000000Z"
                                            },
                                            "currency": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "US Dollar"
                                                    },
                                                    "symbol": {
                                                        "type": "string",
                                                        "example": "USD"
                                                    },
                                                    "exchange_rate": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-12T18:21:39.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-12T18:21:39.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "currency_id",
                                    "tax_cost",
                                    "manufacturing_cost"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Test Create"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description test create."
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1455.1
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "tax_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.01
                                    },
                                    "manufacturing_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 357.66
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 13
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Test Create"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Description test create."
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 1455.1
                                        },
                                        "currency_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "tax_cost": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 15.01
                                        },
                                        "manufacturing_cost": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 357.66
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-12T20:46:05.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-12T20:46:05.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El campo name es obligatorio."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo name es obligatorio."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}